backend/requirements.txt
# Core framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0

# ML and AI
torch==2.1.0
torchvision==0.16.0
transformers==4.36.0
sentence-transformers==2.2.2
chromadb==0.4.18
numpy==1.24.3
scikit-learn==1.3.2

# Image processing
Pillow==10.1.0
opencv-python==4.8.1.78

# Web scraping
playwright==1.40.0
aiohttp==3.9.1
beautifulsoup4==4.12.2
lxml==4.9.3

# Database and caching
redis==5.0.1
sqlalchemy==2.0.23

# Utilities
python-multipart==0.0.6
python-dotenv==1.0.0
psutil==5.9.6
GPUtil==1.4.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1

# Development
black==23.11.0
flake8==6.1.0
isort==5.12.0 